Moved and renamed gridcoordinator to preview
